requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
google-genai
langchain
python-docx
python-multipart